backend/requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
python-multipart==0.0.9
python-dotenv==1.0.1
google-generativeai==0.8.3
groq==0.11.0
openai==1.54.0
easyocr==1.7.1
pillow==10.4.0
numpy>=1.24.0
httpx==0.27.2
sentence-transformers==5.1.2
pdf2image==1.17.0
